Add files via upload
As a new user/institution, I would like to have easy steps with the highest security possible for my assets.

So I have kept the steps really simple and also changed the design aesthetics.

Design Submission includes three user-friendly design pages:-
1. Sign-in with Gnosis - Main Front Page
2. Collective Ownership of NFTs - Collectables Page
3. Emerging Applications Page

1. Sign-in with Gnosis - Main Page
It is always the best experience visually for a user to locate the window easily without too many tabs. 
A simplistic window will open on the first step:
a. New “SAFE” sign-ups
b. Already existing “SAFE”

  2. Collective Ownership of NFTs - Collectables Page
Once the NFT is added, one can view and scroll different collectables. It is always fun to visually see your NFTs and keeps the user interested to keep adding more collectables to the SAFE.

3. Emerging Applications Page
It is best to keep just the logos without too much information about each app on the page. It cleans up the view, and users can easily choose which app to interact with. Placed randomly on the screen and can scroll down for more apps.